+++ src/fractals/littlewood.cpp
@@ -146,7 +146,7 @@ void Littlewood::draw(double) {
     gl::draw_arrays(GL_TRIANGLE_STRIP, 0, 4);
 }
 
-void Littlewood::set(std::string_view parameter, std::optional<float> value) {
+void Littlewood::set(std::string_view parameter, std::optional<double> value) {
     if (parameter == "degree") {
         _degree = GLuint(value.value_or(10));
         add_draw_flag(NEW_DEGREE);
